+++ bashrc/index.d.ts
@@ -1,2 +1,8 @@
 import type { Be } from 'ctx-core/be'
-export declare const bashrc__scp:Be<Promise<void>>
+import type { be_sig_triple_T } from 'ctx-core/rmemo'
+export declare const [
+	bashrc__content$_,
+	bashrc__content_,
+	bashrc__content__set,
+]:be_sig_triple_T<string|false|undefined>
+export declare const bashrc__upload:Be<Promise<void>>