+++ sshd_config/index.d.ts
@@ -1,2 +1,8 @@
 import type { Be } from 'ctx-core/be'
-export declare const sshd_config__scp:Be<Promise<void>>
+import type { be_sig_triple_T } from 'ctx-core/rmemo'
+export declare const [
+	sshd_config__content$_,
+	sshd_config__content_,
+	sshd_config__content__set,
+]:be_sig_triple_T<string|false|undefined>
+export declare const sshd_config__upload:Be<Promise<void>>